Allow body tracking camera to be run at different frame rate than others. Cloud processing for that capture will be disabled and only joints locations will be processed from that camera

+++ src/BodyTrackingUtils.cpp
@@ -12,7 +12,7 @@
 using namespace std;
 using boost::property_tree::ptree;
 
-bool BodyTrackingUtils::predictJoints(ptree* framesJson, int frameCount, k4abt_tracker_t tracker, k4a_capture_t capture_handle, vector<Eigen::RowVector3d>* jointPositions) {
+bool BodyTrackingUtils::predictJoints(ptree* framesJson, double frameCount, k4abt_tracker_t tracker, k4a_capture_t capture_handle, vector<Eigen::RowVector3d>* jointPositions) {
 	k4a_wait_result_t queue_capture_result = k4abt_tracker_enqueue_capture(tracker, capture_handle, K4A_WAIT_INFINITE);
 	if (queue_capture_result != K4A_WAIT_RESULT_SUCCEEDED)
 	{
