+++ src/BodyTrackingUtils.h
@@ -56,7 +56,7 @@ public:
 	* @param jointPositions Container that will hold the predicted joint locations to be used for later processing
 	* @return Whether or not joints could be predicted for the capture frame
 	*/
-	static bool predictJoints(boost::property_tree::ptree* jsonFrames, int frameCount, k4abt_tracker_t tracker, k4a_capture_t capture_handle, std::vector<Eigen::RowVector3d>* jointPositions);
+	static bool predictJoints(boost::property_tree::ptree* jsonFrames, double frameCount, k4abt_tracker_t tracker, k4a_capture_t capture_handle, std::vector<Eigen::RowVector3d>* jointPositions);
 
 	/**
 	* Generates a BoundingBox based on joint locations so that a subject may be isolated within a capture
